Updated a number of plugins to more recent versions.
Note:
A GitHub fork had to be created to create a package source for the credit-card plugin to resolve an issue of it referencing a deprecated property (.title) of the materials text_theme class.

This is because the plugin does not seem to be actively maintained. A pull request with an uploaded fix has been made in any case to the original master.

TODO : Need to review and discuss a future path for a resolution for this.